chore(view) : 파일 네이밍 변경
